requirements.txt
streamlit
scikit-learn
pandas
numpy
PyPDF2
python-docx
joblib
nltk
matplotlib 
fpdf 
google-generativeai